heatmap_fst

Function that generate an Heatmap of Fst and CI values

Usage

heatmap_fst(
  pairwise.fst.full.matrix,
  pairwise.fst.ci.matrix,
  pop.levels = NULL,
  n.s = TRUE,
  digits = 5,
  color.low = "blue",
  color.mid = "yellow",
  color.high = "red",
  text.size = 4,
  plot.size = NULL,
  path.folder = NULL,
  filename = NULL
)

Arguments

pairwise.fst.full.matrix

(object or path).

pairwise.fst.ci.matrix

(object or path).

pop.levels

(optional, character, string) If not supplied, the order is set from the colnames of the full fst matrix. Default: pop.levels = NULL.

n.s

(optional, logical) To have an * when the Fst value is not significant (0 is the lower bound of the CI). Default: n.s = TRUE.

digits

(optional, integer) The number of digits showed in the heatmap. Default: digits = 5.

color.low

(optional, character) Color of lower bound. Default: color.low = "blue".

color.mid

(optional, character) Mid color value. Default: color.mid = "yellow".

color.high

(optional, character) Color of higher bound. Default: color.high = "red".

text.size

(optional, integer) Size of the values. Default: text.size = 2.

plot.size

(optional, integer) By default the size is the number of strata * 2 in cm. Default: plot.size = NULL.

path.folder

(optional, character) Default: path.folder = NULL. Default will use the working directory.

filename

(optional, character) Name of the plot to write. Default: filename = NULL. With default, the plot is not written to disk.